+++ novawallet/Common/Model/LightChainModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct LightChainModel: Decodable, ChainViewModelSource {
+    let chainId: ChainModel.Id
+    let name: String
+    let icon: URL?
+    let options: [LocalChainOptions]?
+}
